+++ src/accessories/index.ts
@@ -1,2 +1,3 @@
+export * from './frame.js';
 export * from './switch.js';
 export * from './television.js';